Don't draw textfields below a certain width
I tracked a pretty serious freezing-issue back to the fact, that
when the window is scaled below the width "imposed" by the size-hints,
a textfield with a width less or equal than 15px would lead to it.

Now, with the different dpi-modes, the margin changes, so I multiply
the 15px with the SCALE-value.
This should also fix related issues on Windows.

This is more like a pretty empiric hack, but to really fix this issue,
one has to discover the reason why the textfields can't be arbitrarily
small.
I'd recommend investigating the PANEL_EDIT-logic.